Move FilterRegistry into filters module.
